Stop testing on node 7 due to unfixed bugs
Node 7 does not include https://github.com/nodejs/node/pull/11015.
Without this fix, the process is killed when proxying a connection
through HTTPS. We could try to work around this on our end, but node 7
is officially unsupported, and it seems more sensible to simply drop
support here entirely too.