Ensure passthrough supports totally transparent proxying
With non-transparent proxying, we're sent the whole URL in the initial GET
path, but with transparent proxying, we're not (the client thinks they're
talking to the real server). We can deal with this, but we need to actively
check & use the Host header, and we need to update the error detection to
ensure we still spot silly errors when using passthrough, rather than
forwarding things endlessly to ourselves.