Make aborted request reporting more reliable
The previous mechanism seems to be broken in Node 10.15.1+.
Unclear exactly why, but could be related to connection
persistence (we check if the response connection is closed,
which presumably may not happen in keep-alive connections).
Fortunately, there's a specific request 'aborted' event
that's supported back to Node 0.3.8, so that will do
nicely.