Create a recordTraffic option, so traffic storage can be disabled
By default (previously and now) each seen request was
attached to the rule it matched, so it could be examined
& verified in testing. This is sometimes useful, but not
always, and uses at least as much memory as the size of
all requests that have ever been seen by the server.

With this change, you can now pass recordTraffic: false,
and traffic data won't be recorded. Note that this does
mean that getSeenRequests will always return an empty
array.