Ignore TSDoc errors - they're its own fault
TSDoc includes its own hardcoded version of TypeScript. It gets
updated occasionally, but not often enough, so it's often out of
date, and will complain about unusual types that depend on the
latest version. We don't really care (as long as its internal
types, not the generated declarations), so let's just ignore
them.