Make request URLs absolute if they are proxied transparently
Transparently proxied requests (where the client doesn't know about
the proxy, i.e. something else has redirected the traffic) will be
received with relative URLs, but should be treated as having
absolute URLs. We now detect this case (via the Host header), and
requests that aren't explicitly sent to the proxy server directly
will now have the absolute URL as their .url property

This may be a breaking change if you are sending requests to a
Mockttp server directly, using a hostname for the server that
isn't 'localhost' or one of its ip addresses. That traffic will
now be treated as having an absolute URL.