Make req.url always absolute, but ensure path matchers match either
This allows for more detail on who the client is talking to when
using Mockttp as a transparent proxy, and significantly more flexible
matching for proxying in general.

This is a breaking change if you examine request.url after calling
getSeenRequests, or in a callback to .on('request', ...). You can
handle it most easily by comparing to server.urlFor('/your-url'),
or you can parse the full URL and compare the path component.