Fix passthrough loop detection bug, partly caused by KA support
Since KA was added, we now hold outgoing ports for long periods of time. This
is great, but passthrough loop checking has a bug: it assumed that ports on the
local machine were unique. In reality, their unique by interface. If we hold
open a port X on the outgoing network interface, and then a local client
connects from port X on the localhost interface, we see it as a loop - in
reality they're independent ports.