Fix bug in protocol-less URL matching
Previously we assumed that parsing of a placeholder
protocol would be the same as parsing HTTP. It's not:
paths don't default to / for example. We now handle
this more rigorously, so URLs without either protocols
or paths will match against root queries correctly,
e.g. example.com will match a request to http://example.com/