Fix cert date new year bug
Previously, if you generated a certificate it set the certificate to start
yesterday, and expire on today's date, but a year later than the *start*
date.

That means that for 2020-1-1, it starts on 2019-31-12 (correct), and
expires on 2020-1-1 (today's date, year after start date), so expires
immediately. This is bad, and would break most cert generation on that
one day. We now use the year after today's date instead.