Avoid double-reporting first-byte client errors
Previously, any error that was visible from the first byte, including
bad methods and a few other cases, was reported twice, seemingly
due to httpolyglot behaviour. This is now worked around: if multiple
client errors fire in the same tick on the same socket, only the
later error will be announced.