Use modern npm, to fix modern node tests
We previously used npm 4, as a lazy way to ensure we didn't use
a package lock anywhere (as a lib, not an application, we want
to always test against the deps an end user will install, not
our preferred versions). Seems like npm 4 (now ancient) is
broken in the last node though, so let's try updating.