Fix error packet peeking with keep-alive requests
Browsers keep-alive aggressively, and this lead to the client-error
header overflow tests showing OGET as the method, when an initial
OPTIONS request was peeked, and then a subsequent GET request was
unparseable.

Didn't come up before, as we didn't use httpolyglot for pure HTTP
serving before.

We now resolve this by only using peeked data for the very first
request, and otherwise just trusting the parser's rawPacket.