Disable TLS-wrapping h2 tests for Node 10, due to client bugs
Not clear exactly why this happens, but there were quite a few HTTP/2
updates & changes in 10.16.3 which could've caused this. It doesn't
happen in Node 12+, or older versions of Node 10, and it only affects
the test client code (the actual server code works fine), so easiest to
just ignore it.