Ignore TLS connection drops if they're not immediate
Where immediate means within 5 seconds. This is useful because
some clients (Chrome) will create pools of connections, and may
not use all connections within the pool. That's not a sign of
any connection or cert-trust issue, it's simply that the it wasn't
necessary, so it's not useful to consider it a TLS error.