Fix bug in HTTP/2 keep-alive handling
Previously we used abort & close events to detect downstream
disconnection. Unfortunately, abort events don't exist, so this actually
just used close events, and with HTTP/2 we get a close event at the end
of each request, not each socket.

When an HTTP/2 connection uses upstream connections it closes and thus
aborts them async, after the request has handled. When using the
keep-alive pool heavily this becomes a problem, as that upstream socket
may now be in use for a different request, so closing one HTTP/2 stream
triggers a different subsequent request to abort entirely. Oops.

We fix this by just listening to the correct 'aborted' event, and only
using it if the response hasn't already completed, to avoid
cross-request interference.