Improve detection & not-warning of TLS preconnects & unused pooling
Previously Chrome especially would create noise, because it drops
connections after a little under 5 seconds, and so still trips the
'rejected with no data' test for TLS connections, even though the socket
is actually fine, and it's just that Chrome preconnected unnecessarily
(probably because a page told it to) or made an extra unused TLS
connection (e.g. because it couldn't assume HTTP/2).

Not clear yet if this covers all cases, but it should be a good
improvement.