Remove now-hopefully-unnecessary SocketWrapper that causes RSTs
We needed this socket wrapper for H1 CONNECTs that contained H2, but it
seems (TBC) that the segfault here is resolved now, since we now handle our
own TLS directly on these connections. With this in place, connection closes
were sometimes done incorrectly when using TLS over the wrapper, resulting
in RST packets after the end of some responses. Didn't break anything, but
lots of noise, and clearly not quite right.