Use pkg.browser, instead of webpack stubbing for browser test deps
Previously webpack (via karma) manually stubbed out node dependencies, a
process that any other users using Mockttp in a browser had to recreate.

Now, we instead mock out non-browser dependencies in package.json's
browser field, which works automatically with webpack etc to do this.
That means we can now remove most (but not quite all...) of the stubbing
in our webpack tests, so we now only stub dependencies of the test code
there, and none of the main module dependencies.