Pin travis to avoid broken Node HTTP/2 client behaviour
This also moves the bar for HTTP/2 TLS over TLS up to 12.17, since it
turns out that's the more precise point it was fixed (relevant, because
the node 12 build is now below it). Not a big deal though - the
functionality does seem to all work fine here, it's purely that the node
client used in testing doesn't.