Change the default standalone port to 45454
This should be an invisible change to end users, as long as you're not
explicitly specifying the standalone server port in browsers tests
on only one side of the client/server setup.

Makes no functional difference, except for ensuring that HTTP Toolkit
itself doesn't have a port conflict with Mockttp tests when both are run
on the same machine.