Add friendly explanations to all endpoint objects
This makes it easier to recognize endpoint objects returned by rule
creation (e.g. thenReply()) or by server.getMockedEndpoints and
server.getPendingEndpoints().

On Node, these objects now console.log as full explanations of their
state. In the browser, this can't be overridden, but there's an
'explanation' field always shown with a clear description of their
behaviour, and toString() will return that on demand.