Don't allow beforeResponse to add custom pseudoheader values
It might actually be interesting to allow this, since it would let
you test some funky behaviour elsewhere, but it's very against the
HTTP/2 spec, and that means that Node won't allow it (and in theory
all clients & intermediaries required to loudly reject it anyway).
Either way, this lets us solve the abiguity of setting statusCode
and :status: only statusCode works, custom :status is rejected.