Make it possible to override :scheme & :authority in beforeRequest
This is useful for some interesting cases where you want to connect
to a server that responds to a hostname, but doesn't resolve as that
hostname. Notably comes up when testing out reverse proxies, CDNs,
etc.

This also adds tests for fallback H1 over H2 proxying, and ensures
that all errors in server request handling are caught and
consistently handled.