Allow injected 0-length bodies on H2 messages that must be empty
Previously, a beforeRequest callback that returned body: '' would result
in an error response for GET/HEAD/DELETE requests over H2. Those methods
aren't allowed bodies, and http2-wrapper enforces this:
https://github.com/szmarczak/http2-wrapper/blob/8e2701887431de4067b5c037c1e4315d94d0d5ee/source/client-request.js#L164

We could hack around that, or try to get that changed, but it'd add
complexity, and it is incorrect behaviour. For now, we ensure that empty
bodies don't cause issues, ignore the intentionally invalid case.