Revert unstable HTTP/2 passthrough loop test
This reverts commit a32a82f2d4182379fa9b6ab28d68488fe7701609.

Unclear why, but this test (more specifically, switching the explanation
tests to HTTPS) makes the handler tests very unstable, during browser
testing specifically. Not clear why, but it's not an especially
important test (functionality already worked before, and logic is
identical to HTTP/1) and it doesn't seem that there's any non-test
issues here, so simplifying tests back to plain HTTP for the moment
seems reasonable.