Always flush upstream request headers ASAP
This only matters if the request body doesn't arrive immediately.
Since the client has sent the request though, we should honor
it's preferences for sending the data on immediately. Note that
this does happen after client body piping, so if the request is
already buffered it should be sent efficiently anyway, I think.