Fix false positives in passthrough loop detection
If you have an open request from the proxy to the proxy that's _not_
being passed through, then you have a local:random -> local:proxyport
request. Previously, we matched the local port, which is a bit
pointless, because a new local:differentrandom -> local:proxyport
request will match that. Not all loopback requests to the proxy port are
the same, and not all are bad.

This got caught because of some probably incorrect HTTP/2 cleanup, which
left a loopback request in the open sockets list during the tests on
Travis somehow. One leftover socket made all future sockets match.

We now match on the random outgoing port, which should be more correct,
and more unique even if it isn't (but it is).