Don't depend on a bad 204 response in tests
This test rewrote a response from 404 to 200, but didn't touch the
headers. The default header included "transfer-encoding: chunked",
which is invalid with the rewritten 204 response.

This is a bad test: really, the client should drop the header if they
want to send that response. This appears to only fail in Node 15+, but
might fail more widely in future if Node is tightening up its HTTP
handling behaviour.