Support full range of types in cors.origin callback
If a server was started with a cors.origin argument, it should enforce
the given origin. The origin can be in a few formats, from string to
regex to callback. The types for the callback here have widened
recently, and this expands the implementation to properly support
all of the possible return values from that callback.