Short-circuit all rule matching once a matching rule is found
This is a performance/non-blocking improvement, it doesn't change the
actual rules for which requests should match.

Previous, as long as it was possible that a rule _might_ match, we
wanted for matcher evaluation to complete before picking the first rule.
We don't need to do that. Instead, we can return as soon as we find
the first rule that matches, and ignore all the others. This doesn't
make much difference most of the time, except when one of the matchers
is waiting for the request body, in which case immediately matching lets
us handle rules that would've matched first regardless slightly quicker.