Revert h2 updates 97edec0 & 8adb5b5 - needs more work and testing
Seems like there's some cases here where this caches some
connections that we don't expect, and that this can cause
Mockttp to miss browser traffic if a server is stopped
and started? Unclear how exactly, but we need to do a
thorough test, and I don't want that to block a release
now.