Improve Node 16 abort handling further
The previous version could lose requests, where they were never
reported with a response or an abort, if the handler failed after
closing the writable response.

In that case, the error handler fires, but fails to write an error, and
so tries to mark the response as aborted. Because the writable is
closed, our previous logic assumed that a response was written OK.
That's not necessarily true. Now, we instead require that all handlers
return ASAP (promises OK, but not IO) after closing the socket, and as
long as they do then everything is OK.

It doesn't seem to be possible to directly reproduce that error case in
the tests here, but the new test at least get closer than the existing
ones.