Very briefly cache passthrough DNS lookups by default
This is useful, because most uses of Mockttp (including Mockttp's own
test suite) repeatedly reference the same hostnames, and Node does
not cache the results. For localhost, this adds 10ms to every new
connection on Linux, which doubles the time taken by most passthrough
tests. DNS is cached only for 10 seconds to ensure that DNS changes
or changes to other mechanisms like Hosts files are quickly reflected
regardless.