Cleanly handle invalid standalone server requests
This used to work, but it looks like this case was missed in the recent
refactorings. Not a huge problem, but it does mean that the server
returns an unclear error for clients that get disconnected (e.g. by a
reset or server restart) and produces a bunch of noisey log output on
the server end.