Add 'mock-server-{started,stopping}' events to the standalone server
This is useful if you want to build automation around the standalone
server itself, e.g. when spawning lots of servers that you want to
log or automatically preconfigure somehow. In HTTP Toolkit, it's
going to allow us to automatically start a Docker-intercepting
proxy alongside every Mockttp proxy server (and clean it up later).