Make @types/node a wildcard dependency to avoid conflicts
For downstream users, the node types used in Mockttp's APIs should
always match the version of node they're using. If we return a Buffer,
then we're returnin a Node vX buffer, where X is up to them.
Conceptually, it's not a fixed dependency we define - it's a property
of their runtime environment, so it makes sense to let them define the
version of the node types used instead of providing our own.

Also, it avoids duplication and conflicts, which otherwise seems
inevitable eventually as users import Mockttp alongside all sorts of
other @types/node versions in future.