Pin types devdependencies
We don't want a package-lock, because it won't reflect the versions
really used by users installing the package, and in general we can't pin
runtime dependencies or we'd get into all sorts of conflicts &
duplication unnecessarily.

Still, it's useful to have some reliability, and types packages
particularly that's important because they don't follow semver: patch
version updates can break the build here. There's no real downside to
pinning devdep types like this, and it avoids that issue, so let's do
it.