Fix issue with H2 over H2 in Node v16.7+
Previously when an HTTP/2 connection was shut down in some tests
it could end up destroying the lower level sockets before the
upper HTTP/2 session tries to write GOAWAY frame to those sockets
which broken everything. This only happens in Node v16.7+ but its
not clear why. It's reasonable to set allowHalfOpen anyway though
to delegate the stream shutdown to the upper layers, so in that
one case we now do so.