Fix serialization channel dispose
Before this, dispose didn't actually work, and was instead raised
as real data in each handler. That didn't trigger as a request or
response, but did count as relevant data, which activated stream
handlers, which broke a test that used a "new Readable()". This
is a bug at the end of the day, since that's irrelevant data at
best and it blocked real dispose - it's now silenced.