Fix keepalive - broken in node by objectMode on client WS stream
WebSockets can never be in objectMode: they're wrappers on sockets,
which only transact in raw buffers or strings.

Without this change, when keepAlive was enabled, the client websocket
would fail to send a pong() response, complaining that the value written
somewhere was the wrong type and crashing the process.

This was in place since it's also used by websocket-stream to disable
multi-chunk websocket messages. We already deal with those though (in
serialization's ClientServerChannel we split messages on newlines) so
that's not actually a problem in our case, as far as I can tell.