Don't rethrow errors from getSeenRequests() after failed requests
Previously, if request handling threw an error (e.g. if the response
was aborted or passthrough request failed entirely) then the error
was stored instead of the seen request, and getSeenRequests would
rethrow that same error later on when queried. This was very
confusing and definitely wrong - the request was seen perfectly
correctly, in general, it's only the response which failed.

We now store the real seen request in all cases, even if the handler
completely fails somehow. To capture errors, it's generally better to
use on('abort') or on('response') instead - but this shouldn't matter
for normal testing use cases.