Add .forX() methods for rule building, deprecating .get, .post etc
The original methods are very widely used, so will continue to be
supported fully at least until the next major version, but the forX
prefix is a bit clearer. The fact that server.get() doesn't actually
send an HTTP request has been a source of significant confusion in
the past. In addition, this helps consistency when supportuing more
methods in the various Mockttp derivatives that are coming soon for
IPFS/WebRTC/Ethereum/etc whilst avoiding any risk of unnecessary
naming conflicts.