Rename 'standalone' servers to 'admin' servers, deprecating old APIs
This is not a breaking change - all existing APIs referencing
standalone servers, such as getStandalone and resetStandalone are
preserved, but deprecated. These will be removed in a future breaking
v3 release, but not imminently.

This change is intended to support some future integrations with the
admin server to allow protocol plugins that share the same admin
infrastructure, and the 'standalone' name has been confusing for a good
while so this seems like a sensible time to bite the bullet and clear
things up a little.