Fix resetAdminServer + client.stop() race condition
Previously, if resetAdminServer was called, there was a possible race
condition that could break tests which waited on a client.stop() call.

This could happen because when the admin server was reset, it closed the
websocket and triggered a client stop(). This correctly shut down the
client, but if it happened right before a real stop() call, then second
stop call would be skipped, and would not wait for the in-progress
stop(), instead returning immediately, implying the server was
immediately correctly shut down, when in fact shutdown was still in
progress.

The client now tracks an internal promise for pending startup &
shutdown, and waits for those to complete if they are in progress when
start() or stop() are called.