BREAKING: Auto-encode beforeRequest/Response callbacks body data
This is almost always what you want: if you return a body in either
beforeRequest or beforeResponse, and you either add or just didn't
remove a content-encoding header, then the body should be encoded
automatically to match that header. For example, if you're modifying a
gzipped response, the body needs to be re-gzipped so that the client
receives it and ungzips it to get your modified data.

A new 'rawBody' field will be added shortly to allow opting in to the
old behaviour, for cases where you want to handle encoding yourself, or
to intentionally break the encoding for testing.