Fix a bug that normalized all headers for remote passthrough rules
Rules were serialized & deserialized such that transformReq/Res was
always an object, even if it was initially unset, which means the
normalization for that was always run. That step also always normalized
headers even when nothing was actually transformed.

In practice, this meant with remotely defined passthrough rules, header
names were all lowercased & deduped, rather than preserving the initial
raw state.

This is now improved somewhat: we serialized & deserialize to avoid
this, and we add some checks to avoid normalizing in a couple of cases.

In future we can probably best improve this by supporting rawHeaders in
transforms & callbacks, but that's a non-breaking change we can ignore
for now.