Update test PFX file format to support Node 18
When using Node 18, OpenSSL v3 is used. This disables many old insecure
algorithms, including RC2, which our test PFX file used. We don't care
about the security of the test file really, but it's useful to have
something that can be loaded for testing in Node 18.

This regenerates the PFX from our existing CA & key files. The openssl
command for this is:

openssl pkcs12 -inkey test-ca.key -in test-ca.pem -export -out test-ca.pfx -descert

This is 3DES which isn't the most up to date cipher we could be using
so we may hit the same problem in future, but that's OK and this serves
as a nice early warning for breakage that might affect users with old
key files too.