Stop forwarding the proxy-connection header, to avoid traffic blocking
In theory we shouldn't be forwarding this regardless, since the RFC
specifically tells us not to ever do that, so this is the more 'correct'
behaviour. That's not really the intent though: we're changing this
because it looks weird from the server-side, and weird enough that many
anti-bot services including Cloudflare will block traffic outright if it
includes this header.

This combined with the previous Firefox-matching TLS changes mean that
we can now avoid all traffic blocks I've been testing against (although
some cases specifically require either HTTP/1 or HTTP/2 to work
correctly).

On tricky case: when proxying with HTTP/2 to an HTTP/1 server. HTTP/2
never preserves header case, but HTTP/1 does (even though it's
officially insensitive). That means all incoming HTTP/2 is lowercase,
but when we fall back to HTTP/1 upstream, we're now sending unexpectedly
lowercase header names, and that's considered suspicious enough to
trigger some blocks. As a workaround users can disable HTTP/2 for those
cases, but not sure what else we can do, since we accept the incoming
HTTP/2 connection before we connect upstream in the first place...