Don't replace body when using matchReplaceBody if there were no matches
This mitigates a potential corruption issue, if data can't be accurately
represented in UTF-8 for some reason. Previously, the data would be
translated into UTF-8 regardless, and then translated back to a raw
buffer and replaced in the response, even if it was the same as the
original response body, which was a lossy operation in these cases.

Now we avoid the potentially lossy operation, unless some part of
the body really does match one of the given strings.

There's still a tiny risk of corruption if there is a real match, but
that might be acceptable (sometimes you really want to replace strings
deep within binary data) and not all transformations of binary data
really are lossy anyway.