Fix bug where including auth params in a proxy URL broke all connections
This happened because http-proxy-agent attempts to make a single
(non-CONNECT) proxy request, including proxy-auth there, and that
resulted in a setHeader() call on the request after the initial request
had been created with raw headers (which permanently sets the headers),
which threw HTTP_HEADERS_ALREADY_SENT.

To fix this, we now drop proxy-agent to implement our own proxy logic,
and always use https-proxy-agent for HTTP(S), so we *always* do a
CONNECT in these cases. This allows the proxy-agent to set the header
for the CONNECT independently of the raw headers set later on. It also
lets us implement our own stricter caching for proxy agents, and it's
going to let us doing better custom configuration for proxy-agents in
other ways too (imminently).