Omit invalid upstream response headers during proxying
This is only really relevant when --insecure-http-parser is set. That
may be useful in some cases when communicating with upstream servers
that intentionally (on uncontrollably) send invalid headers in
responses. Previously, that would result in the entire request failing,
which is unhelpful, and different to browsers (which omit the invalid
header). At least one CDN (Incapsula, now called Imperva) intentionally
uses this to try to block bots.