Replace ja3er.com in tests with ja3.zone, which seems far more reliable
Ja3er has been broken for a fair while, and creates timeouts which we
ignore (to avoid flaky tests) but slow down the testing flow and it's a
bit of a waste. Ja3.zone seems new & solid, looks like a far better
alternative that'll actually let this test run most of the time.