Move to separate socket-reset error UX from reset logic
We can now independently a) require socket-reset support and clearly
error if it's not available and b) aggressively shut down a socket in
the most errory way possible for the remote client. Combining the two
was awkward and duplicative in the past, and this gives us more
flexibility to fix some reset issues shortly.