Fix handling of intentional resets on HTTP/2 & tunnels
socket.resetAndDestroy fails in any cases where the socket handle is not
actually a native TCP socket (e.g. sockets on top of tunnels) or where
the socket is actually an HTTP/2 fake socket.

We now handle this by unwrapping as much as possible, and then sending
explicit RST_FRAME errors for HTTP/2 instead of using TCP reset. This
should get pretty close to representing a connection-level issue in each
case without creating unnecessary breakage elsewhere (e.g. resetting the
TCP socket used by many different proxied connections, when proxying
over HTTP/2).

Note that the HTTP/1-over-HTTP/2 case fails here, due to an ECANCELLED
error in the JS stream wrap layer. This seems to only affect late v8 but
not later, and it's a test client issue, not a Mockttp issue, so just
skipping that test for now.