Ensure body streams resolve even if they've already completed
Only known example where this happens is in HTTP Toolkit, where a rule
sets a response. Best guess that the response completes so fast that
the request is closed before the tick where we attempt to look at it,
and so we're just looking at a completed empty stream by this point.
It's totally fine and legitimate for the body stream to close
immediately, we just need to check for and handle that explicitly.

This fix will only work in Node versions that support readableEnded,
but that's v12.9+ and we officially only support v14+ anyway.