Allow subscribing to 'rule-event', to monitor rule processing events
The first and main motivating example here is upstream passthrough
events. This allows you to see exactly what a passthrough rule is send &
receiving from the server, not just what the client is sending &
receiving to Mockttp in general.

This can be useful for low level debugging of proxy traffic and complex
processing or transformations, where the forwarded traffic is
confusingly different to the original input/resulting output.

This does not attempt to do the same for websocket passthrough yet, as
websocket passthrough doesn't support very much in terms of traffic
modification, but this will be added in future as that develops.

This required touching on a few details of how passthrough worked for
server response bodies, and _may_ have fixed some related bugs there
where the body could be read, not modified, but then lost and not
forwarded back to clients.