Correctly mirror upstream websocket subprotocols back to clients
Previously the first subprotocol was always used automatically,
regardless of the upstream protocol. That was almost always what you
wanted (most cases just use one subprotocol anyway) but could easily
be wrong in some advanced cases. We now specifically match what the
upstream server sends, when forwarding.

When not forwarding, we stick to the same behaviour, replying with the
first suggested protocol, if one is provided.