Add workarounds to fudge & accept mildly invalid WebSocket protocols
Without this, WS will reject invalid subprotocols, for empty header
values, or some syntactically incorrect values. Some real clients send
these headers in ways that are accepted, so it's very useful to be able
to intercept them, and the intent is clear: an empty-string value should
be a null not-set protocol (and is generally treated as such).

In a perfect world, we'd forward the subprotocol exactly as-is without
this, and avoid taking an opinion entirely as long as we could do basic
handling, but WS has opinions and ties our hands on that.