Improve client error lost-byte handling for new Node
In Node 20.10.0, the buffer reported in rawPacket is the same as
previous versions, but it's not === the same as the peeked data exposed
by httpolyglot. This is unfortunate! Something must be rewrapping it
somewhere. To handle this, we now just check for equality.

This feels a little risky, like data could be lost in these errors given
duplicate characters, but I think given the behaviour in practice (where
one char is reported, and then a whole big chunk) it's unlikely to
be possible to hit that case in reality. The test has been slightly
tweaked to double check that case, and I can't find any similar way to
actually trigger this through any other route (just 'QQ' for example).