Fix bug where duplicate headers were lost in passthrough when using H2
Previously, if a client connected with HTTP/2, a request was proxied,
and the response contained duplicated headers (e.g. multiple set-cookie
headers) then all but the last header value was lost. This is due to a
bug in Node's HTTP/2 compatability layer, which will be fixed
separately. In the meantime, this patch should let us work around the
issue with minimal impact.

This patch works by collapsing raw headers back to an object. The only
situation where this will reorder or lose header info is for duplicate
headers, which are conveniently broken by this bug already. Other raw
header cases like key-casing don't matter for H2 because it enforces
lowercase keys regardless.