Avoid issues with port contention on parallel server startup
This did tend to always resolve OK eventually, but seems to cause
intermittent test failures recently in the rule-events test (which
starts 2 servers in parallel). Bizarrely, this seems to result in much
more contention than starting 100 servers in parallel (in
port-configuration).

Either way, this is now resolved. This slightly reduces startup
performance when starting 100s of servers at once, but speeds it up
significantly when starting just a couple (which seems much more
like normal practice).