Fix support for TLSv1 via ignoreHostHttpsErrors on modern Node
Previously we'd assumed that the OpenSSL v3 upgrade in modern node made
this unavailable without command line flags. In fact, it is possible to
set the security level on a per-context basis, so we can still use this
to access old TLS servers compatibly!

This changes adds that, and completely removes the test's compatibility
check & skip for this feature.