Fix unnecessary admin server shutdown due to brief disconnections
Without this, when the device goes to sleep, in many cases the server
stops, and the client disconnects, then attempts and fails to reconnect,
before the machine restarts, resulting in full disconnection and a
self-shutdown of the whole session. This causes problems in HTTP Toolkit
on Windows, where it can be triggered easily in normal usage.

We now retry multiple times, with a brief delay after the first retry.