Normalize IPs during localhost loopback proxying
This shouldn't be visible externally - it just affects internal address
resolution logic.

This is useful because of cacheable-lookup#85 - in Node v20.12+ IPv6 IP
DNS resolution can fail in unusual ways, so we'd rather preserve IPv4
traffic in simple IPv4 format where possible.

The 'better' solution for this would be cacheable-lookup correctly
resolving this address as dns.lookup does, or some kind of wrapper that
avoids resolving IPs entirely (it's just inefficient and asking for
trouble really).