Fix error & head handling in stream handlers
This ensures that errors result in a client response failure, and that
headers are flushed without waiting for the stream to start.

This approach is a little basic, and errors don't yet handle this case
of browser streams, but I think those are very rarely used and it's
unlikely to be relevant.