Update destroyable server to v1.1.1 to avoid races with server.stop()
Without this, there is a race condition given an open connection to a
Mockttp server after calling server.stop() and then server.start() in
rapid succession. It's possible in this case that the client may try to
reuse the connection because server.stop() did not wait until the
connection was fully closed, which means the client may not have
received the FIN on their connection. In rapid localhost testing
scenarios, this can result in failures in the subsequent tests with
some clients.